+++ crates/agent-otel-daemon/src/platforms/gemini.rs
@@ -21,7 +21,7 @@ impl PlatformDescriptor for GeminiQuotaProvider {
     fn aliases(&self) -> &'static [&'static str] {
         AntigravityDescriptor.aliases()
     }
-    fn wire_client_id(&self) -> u8 {
+    fn wire_client_id(&self) -> u16 {
         AntigravityDescriptor.wire_client_id()
     }
     fn pre_tool_response(&self) -> agent_otel_core::platform::HookResponse {
